refactor: 整合CommandZookeeper
