+++ motan-demo/motan-demo-server/src/main/resources/motan_demo_server_commandRegistry.xml
@@ -31,15 +31,18 @@
     <motan:protocol id="demoMotan" default="true" name="motan"/>
 
     <!-- 通用配置，多个rpc服务使用相同的基础配置. group和module定义具体的服务池。export格式为“protocol id:提供服务的端口”-->
-    <motan:basicService id="serviceBasicConfig" group="yf-demo-rpc" registry="registry"
+    <motan:basicService id="serviceBasicConfig" registry="registry"
                         export="demoMotan:8002" accessLog="false"/>
 
     <!-- 具体rpc服务配置，声明实现的接口类。-->
-    <motan:service interface="com.weibo.motan.demo.service.MotanDemoService"
+    <motan:service interface="com.weibo.motan.demo.service.MotanDemoService" group="aaa"
                    ref="motanDemoServiceImpl" export="demoMotan:8001" basicService="serviceBasicConfig">
     </motan:service>
-    <motan:service interface="com.weibo.motan.demo.service.MotanDemoService" group="tc-demo-rpc"
+    <motan:service interface="com.weibo.motan.demo.service.MotanDemoService" group="bbb"
                    ref="motanDemoServiceImpl" export="demoMotan:8002" basicService="serviceBasicConfig">
+    </motan:service>
+    <motan:service interface="com.weibo.motan.demo.service.MotanDemoService" group="ccc"
+                   ref="motanDemoServiceImpl" export="demoMotan:8003" basicService="serviceBasicConfig">
     </motan:service>
 
 </beans>
