--- motan-registry-command-zookeeper/src/main/java/com/weibo/api/motan/registry/zookeeper/ZkNodeType.java
@@ -1,7 +1,0 @@
-package com.weibo.api.motan.registry.zookeeper;
-
-enum ZkNodeType {
-    AVAILABLE_SERVER,
-    UNAVAILABLE_SERVER,
-    CLIENT
-}
