+++ motan-registry-zookeeper/src/main/java/com/weibo/api/motan/registry/ZkNodeType.java
@@ -1,0 +1,7 @@
+package com.weibo.api.motan.registry;
+
+public enum ZkNodeType {
+    AVAILABLE_SERVER,
+    UNAVAILABLE_SERVER,
+    CLIENT
+}
