+++ motan-registry-zookeeper/src/test/java/com/weibo/api/motan/registry/commandZookeeper/CommandZookeeperRegistryTest.java
@@ -1,4 +1,4 @@
-package zookeeper;
+package com.weibo.api.motan.registry.commandZookeeper;
 
 import org.junit.Test;
